+++ packages/experiment-browser/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.16.1](https://github.com/amplitude/experiment-js-client/compare/@amplitude/experiment-js-client@1.16.0...@amplitude/experiment-js-client@1.16.1) (2025-07-01)
+
+
+### Bug Fixes
+
+* url-redirect impression events fired on redirected page ([#193](https://github.com/amplitude/experiment-js-client/issues/193)) ([a80ac61](https://github.com/amplitude/experiment-js-client/commit/a80ac612290c065f0a4ab877c2f02c2ba0c227c6))
+
+
+
+
+
+# [1.16.0](https://github.com/amplitude/experiment-js-client/compare/@amplitude/experiment-js-client@1.15.6...@amplitude/experiment-js-client@1.16.0) (2025-06-23)
+
+
+### Features
+
+* add throwOnError configuration option ([#187](https://github.com/amplitude/experiment-js-client/issues/187)) ([00761aa](https://github.com/amplitude/experiment-js-client/commit/00761aa122c4f5da30af914a82543635bfd25208))
+
+
+
+
+
 ## [1.15.6](https://github.com/amplitude/experiment-js-client/compare/@amplitude/experiment-js-client@1.15.5...@amplitude/experiment-js-client@1.15.6) (2025-05-06)